+++ src/Network/ArticleService.js
@@ -1,4 +1,5 @@
 import * as API from './APIType';
+import { useQuery } from 'react-query';
 
 const articleUrl = path => {
   return `${API.url('/articles')}${path}`;
@@ -145,16 +146,23 @@ const ArticleService = {
    * `200` : success
    * `401` : fail
    */
-  getArticleById: async articlesId => {
+  getArticleById: async articleId => {
     const method = 'GET';
-    const url = articleUrl(`/${articlesId}`);
+    const url = articleUrl(`/${articleId}`);
 
-    const res = await API.AXIOS({
+    const response = await API.AXIOS({
       method,
       url,
     });
-    return res.data;
+    return response.data;
   },
+  // useArticle(articleId) {
+  //   return useQuery(
+  //     ['getArticleById', articleId],
+  //     async () => await this.getArticleById(articleId),
+  //     { suspense: true },
+  //   );
+  // },
   /**
    * **UPDATE** One Articles By Articles ID
    * @param {string} articlesId
@@ -233,18 +241,25 @@ const ArticleService = {
     const url = articleUrl(`/${articlesId}/comments`);
     const params = { order, page, take };
 
-    let response;
-    try {
-      response = await API.AXIOS({
-        method,
-        url,
-        params,
-      });
-    } catch (error) {
-      alert(error);
-    }
+    const response = await API.AXIOS({
+      method,
+      url,
+      params,
+    });
     return response.data;
   },
+  // useComments(articleId, order, page, take) {
+  //   return useQuery(
+  //     ['getCommentsById'],
+  //     async () =>
+  //       await ArticleService.getArticlesCommentsById(
+  //         articleId,
+  //         order,
+  //         page,
+  //         take,
+  //       ),
+  //   );
+  // },
   editArticles: async (articlesId, articles) => {
     const method = 'PUT';
     const url = articleUrl(`/${articlesId}`);
