+++ src/Pages/ArticlePage/Components/Comment.jsx
@@ -1,70 +1,88 @@
-import { useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
+import { useQuery, useMutation, useQueryClient } from 'react-query';
 
-import { getProfile } from 'Utils';
-import { getArticleTime } from 'Utils/dayjsUtils';
-import { ReactionService, CommentService } from 'Network';
+import { ArticleService, CommentService } from 'Network';
 
-import { FavoriteBorder } from '@mui/icons-material';
-import FavoriteIcon from '@mui/icons-material/Favorite';
+import useInput from './useInput';
+import { useArticle, useComments, useCreateComment } from './hooks';
+import CommentView from './CommentView';
+import Fab from '@mui/material/Fab';
+import ArrowUpwardRoundedIcon from '@mui/icons-material/ArrowUpwardRounded';
+import { SmsOutlined } from '@mui/icons-material';
 
 import Styled from '../ArticlePage.styled';
-const Comment = ({
-  curUser,
-  articleId,
-  comment,
-  isLikeInitial,
-  likeCountInitial,
-  onDeleteComment,
-}) => {
-  const [isLike, setIsLike] = useState(isLikeInitial);
-  const [likeCount, setLikeCount] = useState(likeCountInitial);
 
-  const handleClickLike = async id => {
-    const res = await ReactionService.createCommentReactionHeart(articleId, id);
+const Comment = ({ articleId, currentUser }) => {
+  const lastComment = useRef();
+  const queryClient = useQueryClient();
+  const [input, handleChange, reset] = useInput();
+  const {
+    data: { categoryId },
+  } = useArticle(articleId);
+  const {
+    data: { meta, data },
+  } = useComments(articleId, 'ASC', 1, 1000);
+  const likeComment = useMutation();
+  const deleteComment = useMutation();
+  const createComment = useCreateComment(input, articleId, lastComment);
+  const [comments, setComments] = useState([]);
+  const [totalCount, setTotalCount] = useState();
 
-    setIsLike(res.isLike);
-    setLikeCount(res.likeCount);
+  const handleDeleteComment = commentId => {
+    setComments(prev => prev.filter(comment => comment.id !== commentId));
+    fetchComment();
+  };
+  const fetchComment = async () => {
+    const res = await ArticleService.getArticlesCommentsById(
+      articleId,
+      'ASC',
+      1,
+      1000, // 한번에 1000개 긁어옴. 어떻게 할 지 결정 후 바꿔야 함.
+    );
+    setComments(res?.data || []);
+    setTotalCount(meta.totalCount);
   };
 
-  const handleClickDelete = async commentId => {
-    await CommentService.deleteComments(commentId);
-    onDeleteComment(commentId);
+  const handleClickSubmit = e => {
+    e.preventDefault();
+    createComment.mutate();
+    reset();
   };
   return (
-    <Styled.CommentDiv isMine={curUser.id === comment.writer.id}>
-      <div className="info">
-        <Styled.ProfileImage
-          width="2.4rem"
-          src={getProfile.findProfileById(comment.writer.character)}
-        />
-        <div className="picture"></div>
-        <div className="text">
-          <h1>{comment.writer.nickname}</h1>
-          <h2>{getArticleTime(comment.updatedAt)}</h2>
-        </div>
-      </div>
-      <Styled.CommentContent
-        liked_count={likeCount}
-        isMine={curUser.id === comment.writer.id}
+    <div className="comment_list_div">
+      <Styled.ArticleCommentDiv
+        className="comment_count"
+        commentCount={meta.totalCount}
       >
-        <div className="text">{comment.content}</div>
-        {curUser.id === comment.writer.id ? (
-          <button
-            className="delete_button"
-            onClick={() => handleClickDelete(comment.id)}
-          >
-            삭제
-          </button>
-        ) : (
-          <span
-            className="liked_count"
-            onClick={() => handleClickLike(comment.id)}
-          >
-            {isLike ? <FavoriteIcon /> : <FavoriteBorder />}
-          </span>
-        )}
-      </Styled.CommentContent>
-    </Styled.CommentDiv>
+        <SmsOutlined />
+      </Styled.ArticleCommentDiv>
+      {categoryId !== 3 &&
+        data.map(comment => (
+          <CommentView
+            curUser={currentUser}
+            key={comment.id}
+            articleId={articleId}
+            comment={comment}
+            isLikeInitial={comment.isLike}
+            likeCountInitial={comment.likeCount}
+            onDeleteComment={handleDeleteComment}
+          />
+        ))}
+      <Styled.CreateCommentDiv>
+        <form onSubmit={handleClickSubmit}>
+          <input
+            required
+            value={input}
+            onChange={handleChange}
+            placeholder="댓글을 입력하세요"
+          />
+          <Fab className="fab_button" type="submit">
+            <ArrowUpwardRoundedIcon />
+          </Fab>
+        </form>
+      </Styled.CreateCommentDiv>
+      <div ref={lastComment}></div>
+    </div>
   );
 };
